factored out example plotting code, refactored ising and grbm examples
