+++ examples/example_mnist_grbm.py
@@ -8,7 +8,7 @@ from paysage import fit
 from paysage import optimizers
 from paysage import backends as be
 
-from helper import default_paths, example_plot
+import example_util as util
 import plotting
 
 transform = partial(batch.scale, denominator=255)
@@ -21,7 +21,8 @@ def example_mnist_grbm(paysage_path = None, show_plot = True):
     learning_rate = 0.001
     mc_steps = 1
 
-    (paysage_path, filepath, shuffled_filepath) = default_paths(paysage_path)
+    (paysage_path, filepath, shuffled_filepath) = \
+        util.default_paths(paysage_path)
 
     # set up the reader to get minibatches
     data = batch.Batch(shuffled_filepath,
@@ -63,41 +64,14 @@ def example_mnist_grbm(paysage_path = None, show_plot = True):
     # evaluate the model
     # this will be the same as the final epoch results
     # it is repeated here to be consistent with the sklearn rbm example
-    performance = fit.ProgressMonitor(0,
-                                      data,
-                                      metrics=['ReconstructionError',
-                                               'EnergyDistance',
-                                               'EnergyGap',
-                                               'EnergyZscore'])
-    print('Final performance metrics:')
-    performance.check_progress(rbm, 0, show=True)
+    metrics = ['ReconstructionError', 'EnergyDistance',
+               'EnergyGap', 'EnergyZscore']
+    performance = fit.ProgressMonitor(0, data, metrics=metrics)
 
-    print("\nPlot a random sample of reconstructions")
-    v_data = data.get('validate')
-    sampler = fit.DrivenSequentialMC(rbm)
-    sampler.set_state(v_data)
-    sampler.update_state(1)
-    v_model = rbm.deterministic_step(sampler.state)
-
-    idx = numpy.random.choice(range(len(v_model)), 5, replace=False)
-    grid = numpy.array([[v_data[i], v_model[i]] for i in idx])
-    example_plot(grid, show_plot)
-
-    print("\nPlot a random sample of fantasy particles")
-    random_samples = rbm.random(v_data)
-    sampler = fit.DrivenSequentialMC(rbm)
-    sampler.set_state(random_samples)
-    sampler.update_state(1000)
-    v_model = rbm.deterministic_step(sampler.state)
-
-    idx = numpy.random.choice(range(len(v_model)), 5, replace=False)
-    grid = numpy.array([[v_model[i]] for i in idx])
-    example_plot(grid, show_plot)
-
-    print("\nPlot a random sample of the weights")
-    idx = numpy.random.choice(range(be.shape(rbm.params['weights'])[1]), 5, replace=False)
-    grid = numpy.array([[rbm.params['weights'][:, i]] for i in idx])
-    example_plot(grid, show_plot)
+    util.show_metrics(rbm, performance)
+    util.show_reconstructions(rbm, data.get('validate'), fit, show_plot)
+    util.show_fantasy_particles(rbm, data.get('validate'), fit, show_plot)
+    util.show_weights()
 
     # close the HDF5 store
     data.close()
