+++ examples/example_mnist_ising.py
@@ -1,17 +1,16 @@
-import os, sys, numpy, pandas
-
 from paysage import batch
 from paysage.models import visible
 from paysage import fit
 
-from helper import default_paths, example_plot
+import example_util as util
 import plotting
 
 def example_mnist_ising(paysage_path=None, show_plot = False):
 
     batch_size = 50
 
-    (paysage_path, filepath, shuffled_filepath) = default_paths(paysage_path)
+    (paysage_path, filepath, shuffled_filepath) = \
+        util.default_paths(paysage_path)
 
     # set up the reader to get minibatches
     data = batch.Batch(shuffled_filepath,
@@ -27,28 +26,17 @@ def example_mnist_ising(paysage_path=None, show_plot = False):
     ising.initialize(data, method='tap')
 
     # evaluate the model
-    performance = fit.ProgressMonitor(0,
-                                      data,
-                                      metrics=['EnergyDistance',
-                                               'EnergyGap',
-                                               'EnergyZscore'])
-    print('Final performance metrics:')
-    performance.check_progress(ising, 0, show=True)
+    metrics = ['EnergyDistance','EnergyGap', 'EnergyZscore']
+    performance = fit.ProgressMonitor(0, data, metrics=metrics)
 
-    print("\nPlot a random sample of fantasy particles")
-    v_data = data.get('validate')
-    random_samples = ising.random(v_data)
-    sampler = fit.DrivenSequentialMC(ising)
-    sampler.set_state(random_samples)
-    sampler.update_state(1000)
-    v_model = ising.deterministic_step(sampler.state)
-
-    idx = numpy.random.choice(range(len(v_model)), 5, replace=False)
-    grid = numpy.array([[v_model[i]] for i in idx])
-    example_plot(grid, show_plot)
+    # check model
+    util.show_metrics(performance, ising)
+    util.show_fantasy_particles(ising, data.get('validate'), fit, show_plot)
 
     # close the HDF5 store
     data.close()
+    print("Done")
+
 
 if __name__ == "__main__":
     example_mnist_ising(show_plot = False)
